Added suggestions and questions
I added my suggestions and questions as comments first so that we can discuss what to do on it. Please dont add anything to our code, for now lets just add comments when you have ideas and we should discuss our ideas so that we wont have misunderstandings and so that our codes will not overlap and it'd be easier to debug this way too. It won't take too long to code it when everyone is on the same page, and we can distribute the work properly like this.