Add files via upload
ERRORS: inside BTManager and NodeManager
Why?
1.  When the root node is split, it doesnt change the node or make new nodes
2. Every value is correct except for the root that was split, the one in record node #3
What I added
GetAllChildren in BTManager
So that the children can be sorted together with the keys and offsets.
Parameters in addNode, changeNode so that the values of the children can be passed, this is so that whenever a parent node is split, and addNode and changeNode are called on this node, the children wont be replaced by values -1.
GetAllOffsetOfParent GetAllKeysOfParent GetAllChildrenOfParent inside NodeManager
This is so that the recursion of sort will update to the values of the parent
Split:
I turned sort into a recursive method that will check for the parent and insert accordingly
the base case is when  CURRENT NODE TO BE SPLIT IS ALREADY THE ROOT NODE
Why? Because if the parent node has space, this means that there is no need to check the parent of that parent and so on. If the current node to be split is already the root node then this must changeNode, addNode, and placeParent .

WHERES THE ERROR?
Probably somewhere in placeParent
and the if else statement in split doesnt work properly

DOUBLE CHECK THESE MEHTODS:
insert2ChildorNot (i wasnt able to review this that much)
split (the if else statement, maybe parameters are wrong during recursion)
placeParent (


As far as I can remember these are the stuff I was able to do, please just tell me if you need to ask something and my logic is not always right so please check them too!